agent_worker/requirements.txt
browser-use==0.6.1
playwright
pydantic
python-dotenv
openai
aiofiles>=24.1.0
httpx